# sql: don't allocate createdSequences map on each transaction