Previously, `connExecutor.resetExtraTxnState` would re-allocate the `extraTxnState.createdSequences` hash map on each transaction. This commit changes the map to be lazily allocated on the transaction's first call to `addCreatedSequence`. This avoids the allocation in the common case.

```sh
➜ benchdiff --run='BenchmarkKV/./SQL/rows=1$$' --count=25 ./pkg/sql/tests

name                     old time/op    new time/op    delta
KV/Update/SQL/rows=1-10     267µs ± 3%     263µs ± 2%  -1.42%  (p=0.000 n=25+21)
KV/Delete/SQL/rows=1-10     181µs ± 1%     180µs ± 1%  -0.56%  (p=0.003 n=24+24)
KV/Scan/SQL/rows=1-10       113µs ± 2%     113µs ± 1%  -0.41%  (p=0.049 n=22+23)
KV/Insert/SQL/rows=1-10     155µs ± 1%     155µs ± 2%  -0.32%  (p=0.045 n=23+23)

name                     old alloc/op   new alloc/op   delta
KV/Scan/SQL/rows=1-10      31.8kB ± 0%    31.7kB ± 0%  -0.22%  (p=0.000 n=25+21)
KV/Insert/SQL/rows=1-10    58.7kB ± 0%    58.7kB ± 1%    ~     (p=0.649 n=25+23)
KV/Update/SQL/rows=1-10    70.3kB ± 1%    70.3kB ± 0%    ~     (p=0.670 n=24+23)
KV/Delete/SQL/rows=1-10    84.3kB ± 0%    84.2kB ± 0%    ~     (p=0.056 n=24+25)

name                     old allocs/op  new allocs/op  delta
KV/Scan/SQL/rows=1-10         359 ± 0%       357 ± 0%  -0.40%  (p=0.000 n=25+23)
KV/Delete/SQL/rows=1-10       603 ± 0%       602 ± 0%  -0.21%  (p=0.000 n=24+25)
KV/Insert/SQL/rows=1-10       485 ± 0%       484 ± 0%  -0.21%  (p=0.000 n=21+18)
KV/Update/SQL/rows=1-10       721 ± 1%       720 ± 1%    ~     (p=0.060 n=23+23)
```

Epic: None
Release note: None